+++ app/policies/flood_risk_engine/enrollment_policy.rb
@@ -27,6 +27,10 @@ module FloodRiskEngine
       (system_user? || super_agent_user?) && enrollment_submitted?
     end
 
+    def change_assisted_digital?
+      (system_user? || super_agent_user?) && enrollment_submitted?
+    end
+
     def edit?
       (system_user? || super_agent_user? || admin_agent_user?) && enrollment_submitted?
     end
